+++ src/test/java/org/osgl/bootstrap/BootstrapTest.java
@@ -1,0 +1,14 @@
+package org.osgl.bootstrap;
+
+import org.junit.Assert;
+import org.junit.Test;
+
+public class BootstrapTest extends Assert {
+
+    @Test
+    public void bootstrapVersionShallContainsBootstrapArtifactId() {
+        Version version = Bootstrap.VERSION;
+        assertEquals("osgl-bootstrap", version.getArtifactId());
+    }
+
+}