+++ src/test/java/org/osgl/bootstrap/VersionTest.java
@@ -24,21 +24,33 @@ import net.evil.pkg.Kit;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
+import org.mockito.ArgumentCaptor;
+import org.mockito.Mockito;
 import org.mrcool.swissknife.SwissKnife;
 import org.mrcool.swissknife.db.DbUtil;
 import org.mrcool.swissknife.internal.StringUtil;
 import org.mrsuck.MyTool;
+import org.slf4j.Logger;
+
+import java.lang.reflect.Field;
 
 public class VersionTest extends Assert {
 
+    protected Logger logger;
+
     @Before
-    public void prepare() {
+    public void prepare() throws Exception {
         Version.clearCache();
+        logger = Mockito.mock(Logger.class);
+        Field field = Version.class.getDeclaredField("logger");
+        field.setAccessible(true);
+        field.set(null, logger);
     }
 
     @Test
     public void itShallLoadVersionInfoFromResourceIfCacheNotHit() {
         Version version = Version.of(SwissKnife.class);
+        assertEquals("swissknife", version.getArtifactId());
         assertEquals("1.0", version.getProjectVersion());
         assertEquals("3a77", version.getBuildNumber());
     }
@@ -56,8 +68,9 @@ public class VersionTest extends Assert {
         Version.of(SwissKnife.class);
 
         // load subpackage version
-        Version version = Version.of(new DbUtil());
-        assertEquals("0.8", version.getProjectVersion());
+        Version version = Version.of(DbUtil.class);
+        assertEquals("swissknife-db", version.getArtifactId());
+        assertEquals("0.8-SNAPSHOT", version.getProjectVersion());
         assertEquals("", version.getBuildNumber());
     }
 
@@ -82,19 +95,60 @@ public class VersionTest extends Assert {
     }
 
     @Test
-    public void itShallPrintVersionAndBuildNumberConcatenatedByDashWhenBuildNumberPresented() {
-        Version version = Version.of(SwissKnife.class);
-        String[] sa = version.toString().split("-");
-        assertEquals(2, sa.length);
-        assertEquals(sa[0], version.getProjectVersion());
-        assertEquals(sa[1], version.getBuildNumber());
+    public void testDecoratedSnapshotProjectVersion() {
+        String projectVersion = "1.0.0-SNAPSHOT";
+        assertEquals("v" + projectVersion, Version.decoratedProjectVersion(projectVersion));
     }
 
     @Test
-    public void itShallOnlyPrintVersionIfBuildNumberIsNotPresented() {
+    public void testDecoratedProjectVersion() {
+        String projectVersion = "1.0.0";
+        assertEquals("R" + projectVersion, Version.decoratedProjectVersion(projectVersion));
+    }
+
+    @Test
+    public void versionTagShallBeCombinationOfDecoratedProjectVersionAndDecoratedBuildNumberIfBuildNumberDefined() {
+        Version version = Version.of(SwissKnife.class);
+        String versionTag = version.getVersion();
+        String expected = Version.decoratedProjectVersion(version.getProjectVersion())
+                + "-" + version.getBuildNumber();
+        assertEquals(expected, versionTag);
+    }
+
+    @Test
+    public void versionTagShallBeDecoratedProjectVersionWhenBuildNumberIsNotDefined() {
         Version version = Version.of(DbUtil.class);
-        String versionTag = version.toString();
-        assertFalse(versionTag.contains("-"));
-        assertEquals(version.getProjectVersion(), versionTag);
+        assertEquals(Version.decoratedProjectVersion(version.getProjectVersion()), version.getVersion());
+    }
+
+    @Test
+    public void itShallPrintArtifactsAndVersionTagInToString() {
+        Version versionWithoutBuildNumber = Version.of(DbUtil.class);
+        String expected = versionWithoutBuildNumber.getArtifactId() + "-" +versionWithoutBuildNumber.getVersion();
+        assertEquals(expected, versionWithoutBuildNumber.toString());
+
+        Version versionWithBuildNumber = Version.of(SwissKnife.class);
+        expected = versionWithBuildNumber.getArtifactId() + "-" +versionWithBuildNumber.getVersion();
+        assertEquals(expected, versionWithBuildNumber.toString());
+    }
+
+    @Test
+    public void itShallReturnUnknownAndLogErrorMessageIfNoArtifactDefinedInVersionFile() {
+        assertSame(Version.UNKNOWN, Version.of("org.demo.badversion.noart"));
+        ArgumentCaptor<String> messageCaptor = ArgumentCaptor.forClass(String.class);
+        ArgumentCaptor<Object> messageArgCaptor = ArgumentCaptor.forClass(Object.class);
+        Mockito.verify(logger, Mockito.times(1)).error(messageCaptor.capture(), messageArgCaptor.capture());
+        assertTrue(messageCaptor.getValue().contains("artifact not defined in .version file"));
+        assertTrue(messageArgCaptor.getValue().toString().equals("org.demo.badversion.noart"));
+    }
+
+    @Test
+    public void itShallReturnUnknownAndLogErrorMessageIfNoVersionDefinedInVersionFile() {
+        assertSame(Version.UNKNOWN, Version.of("org.demo.badversion.noversion"));
+        ArgumentCaptor<String> messageCaptor = ArgumentCaptor.forClass(String.class);
+        ArgumentCaptor<Object> messageArgCaptor = ArgumentCaptor.forClass(Object.class);
+        Mockito.verify(logger, Mockito.times(1)).error(messageCaptor.capture(), messageArgCaptor.capture());
+        assertTrue(messageCaptor.getValue().contains("version not defined in .version file"));
+        assertTrue(messageArgCaptor.getValue().toString().equals("org.demo.badversion.noversion"));
     }
 }