+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+### 1.0.0-BETA-3 (2017-09-05)
+* Introduce `artifactId` into Version data
+* Add `Version.getVersion()` method which returns combination of projectVersion and buildNumber
+* Add `Bootstrap` class with `Version` constant defined
+
 ### 1.0.0-BETA-2 (2017-09-04)
 * Fix `Version.toString()` issue with `buildNumber` be empty string